Push tag if our artifact is of newer source than remote's
Author time is not monotonic, but committer time is.

The assumption is that we arbitrate between two artifacts from the same
source branch but at different revisions. The newest one(*) should win.

(*) Well, the newest one that is not ahead of its remote. Personal local
commits will be ahead but not accessible upstream. This remains TBD.